+++ authorization/src/main/java/org/grnet/status/authorizations/dtos/Attributes.java
@@ -1,0 +1,12 @@
+package org.grnet.status.authorizations.dtos;
+
+import java.util.List;
+
+public class Attributes {
+
+    private List<String> localEntitlements;
+
+    public List<String> getLocalEntitlements() {
+        return localEntitlements;
+    }
+}