+++ authorization/src/main/java/org/grnet/status/authorizations/dtos/GroupUserResponse.java
@@ -1,0 +1,16 @@
+package org.grnet.status.authorizations.dtos;
+
+import java.util.List;
+
+public class GroupUserResponse {
+
+    public String id;
+
+    public String username;
+    public String firstName;
+    public String lastName;
+
+    public String email;
+
+    public List<UserGroupInfoDto> tenants;
+}